MT#55447 introduce proper a=fmtp parsing
Parsing out the a=fmtp string has been left up to the codec init
function until now, with the values that resulted from the parsing being
stored only within the codec. Convert this to an explicit method to
parse the a=fmtp string, and introduce a dedicated struct to store the
resulting values.

Functionally this change is a no-op.

Change-Id: Ia84e26d632ed5209b4439fd82c1e4e38850fd024